requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
requests==2.32.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.1.0
langdetect==1.0.9
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic[email]==2.5.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
